requirements.txt
discord.py
python-dotenv
redis==3.3.11